+++ AUTHORS
@@ -1,0 +1,4 @@
+# This file lists all individuals having contributed content to the repository.
+# For how it is generated, see 'https://github.com/moul/rules.mk'
+
+Manfred Touron <m@42.am>
