+++ go.mod
@@ -11,7 +11,6 @@ require (
 	github.com/golang/protobuf v1.2.0
 	github.com/google/uuid v1.0.0 // indirect
 	github.com/gopherjs/gopherjs v0.0.0-20181103185306-d547d1d9531e // indirect
-	github.com/gorilla/context v1.1.1 // indirect
 	github.com/gorilla/handlers v1.4.0
 	github.com/gorilla/mux v1.6.2
 	github.com/grpc-ecosystem/go-grpc-middleware v1.0.1-0.20190104160321-4832df01553a
@@ -19,7 +18,6 @@ require (
 	github.com/huandu/xstrings v1.2.0 // indirect
 	github.com/imdario/mergo v0.3.6 // indirect
 	github.com/jtolds/gls v4.2.1+incompatible // indirect
-	github.com/moul/advanced-ssh-config v2.8.0+incompatible // indirect
 	github.com/moul/numberinfo v0.0.0-20160903151149-753f372df509
 	github.com/oxtoacart/bpool v0.0.0-20150712133111-4e1c5567d7c2
 	github.com/pkg/errors v0.8.1
@@ -37,9 +35,10 @@ require (
 	golang.org/x/crypto v0.0.0-20181203042331-505ab145d0a9
 	golang.org/x/image v0.0.0-20190118043309-183bebdce1b2 // indirect
 	golang.org/x/net v0.0.0-20190110044637-be1c187aa6c6
-	google.golang.org/appengine v1.2.0 // indirect
 	google.golang.org/genproto v0.0.0-20190108161440-ae2f86662275
 	google.golang.org/grpc v1.17.0
 	ultre.me/kryptos v0.0.0-20181023194748-240fe1a16033
 	ultre.me/recettator v0.4.1-0.20190210231503-241df3a046a3
 )
+
+go 1.13
