+++ pkg/views/funcmap.go
@@ -18,7 +18,6 @@ import (
 	"github.com/disintegration/imaging"
 	"go.uber.org/zap"
 	"golang.org/x/crypto/sha3"
-
 	"ultre.me/calcbiz/pkg/random"
 )
